refactoring : 2단계 - 문자열 덧셈 계산기
리뷰 사항 반영

1. @ParameterizedTest 테스트 어노테이션의 활용

2. 생성 비용이 비싼 객체는 캐싱하여 재사용

3. 객체지향 생활 체조 원칙 규칙 1: 한 메서드에 오직 한 단계의 들여쓰기(indent)만 한다

4. 매직넘버

5. 디렉토리 구조

ref : #3452
