+++ src/main/java/exception/ValidateNegative.java
@@ -1,4 +1,4 @@
-package calculator;
+package exception;
 
 public class ValidateNegative extends RuntimeException {
     public static final String MESSAGE = "음수는 입력할 수 없습니다.";
